CRUD up and working on backend

+++ backend/src/main/java/com/deroahe/gallerybe/service/impl/CommentServiceImpl.java
@@ -1,7 +1,11 @@
 package com.deroahe.gallerybe.service.impl;
 
 import com.deroahe.gallerybe.model.Comment;
+import com.deroahe.gallerybe.model.Image;
+import com.deroahe.gallerybe.model.User;
 import com.deroahe.gallerybe.repository.CommentRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -11,17 +15,23 @@ import java.util.List;
 public class CommentServiceImpl {
 
     CommentRepository commentRepository;
-    UserServiceImpl userService;
-    ImageServiceImpl imageService;
+
+    Logger logger = LoggerFactory.getLogger(CommentServiceImpl.class);
 
     @Autowired CommentServiceImpl(CommentRepository commentRepository, UserServiceImpl userService, ImageServiceImpl imageService) {
         this.commentRepository = commentRepository;
-        this.userService = userService;
-        this.imageService = imageService;
     }
 
     public Comment findCommentById(int id) {
         return commentRepository.findByCommentId(id);
+    }
+
+    public List<Comment> findCommentsByUser(User user) {
+        return commentRepository.findAllByCommentUser(user);
+    }
+
+    public List<Comment> findCommentsByImage(Image image) {
+        return commentRepository.findAllByCommentImage(image);
     }
 
     public List<Comment> findAllComments() {
@@ -32,17 +42,34 @@ public class CommentServiceImpl {
         return commentRepository.save(comment);
     }
 
-    public Comment saveComment(Comment comment, int imageId, int userId) {
-        return commentRepository.save(comment);
-    }
-
     public void saveAllComments(List<Comment> comments) {
         for (Comment comment : comments) {
             saveComment(comment);
         }
     }
 
-    public void deleteCommentById(int id) {
+    public Comment updateComment(Comment comment) {
+        if (!commentRepository.existsById(comment.getCommentId())) {
+            logger.error("Comment id not in DB");
+            return null;
+        }
+        return commentRepository.save(comment);
+    }
+
+    public boolean deleteCommentById(int id) {
+        if (!commentRepository.existsById(id)) {
+            logger.error("Comment id not in DB");
+            return false;
+        }
         commentRepository.deleteByCommentId(id);
+        return true;
+    }
+
+    public void deleteCommentByUser(User user) {
+        commentRepository.deleteByCommentUser(user);
+    }
+
+    public void deleteCommentByImage(Image image) {
+        commentRepository.findAllByCommentImage(image);
     }
 }
